Use webpack when running the tests
This allows us to leverage the use of `require()` on both the tests and
the JS modules, allowing testing on specific modules and avoid a cluttered
global scope.

SugarApi tests aren’t ready for crosstab library to be loaded, thus a hack
was added to keep the tests that support Nomad’s flow. Mango uses the crosstab
flow and this needs to be tested or swapped to service workers.